modularizing code - further minor file separation
